+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - [Improvement] Collect both stored lag and lag.
 - [Improvement] Introduce states and metrics schema validation.
 - [Improvement] Prevent locking in sampler for time of OS data aggregation.
+- [Improvement] Collect and report number of messages in particular jobs.
 - [Fix] Fix a case where states materializing consumer would update state too often.
 - [Fix] Fix a bug when rapid non-initialized shutdown could mess up the metrics.
 - [Fix] Fix a case where upon multiple rebalances, part of the states materialization could be lost.